refactor: 532 - change course-main structure according to FSD
