+++ src/widgets/course-main/helpers/get-course-status.ts
