+++ src/widgets/course-main/helpers/has-day.ts
