+++ src/widgets/course-main/index.ts
@@ -1,1 +1,3 @@
-export { CourseMain } from './course-main';
+export type { CourseMainProps } from './ui/course-main';
+export { CourseMain } from './ui/course-main';
+export { loader } from './api/loader';
