+++ src/widgets/course-main/ui/course-main.module.scss
