+++ src/widgets/course-main/ui/course-main.test.tsx
